// src/pspdfkit.d.ts
type ImageAnnotationJSON = Omit<BaseAnnotationJSON, "type"> & {
  type: "pspdfkit/image";
  description?: string | null;
  fileName?: string | null;
  contentType: string;
  imageAttachmentId: string;
  rotation: number;
  isSignature?: boolean;
  xfdfAppearanceStream?: string;
  xfdfAppearanceStreamOriginalPageRotation?: number;
};
type ShapeAnnotationJSON = Omit<BaseAnnotationJSON, "type"> & {
  strokeWidth: number;
  strokeColor: string | null;
  fillColor: string | null;
  strokeDashArray?: [number, number] | null;
  measurementPrecision?: IMeasurementPrecision | null;
  measurementScale?: MeasurementScaleJSON | null;
  lineWidth?: number | null;
};
type EllipseAnnotationJSON = ShapeAnnotationJSON & {
  type: "pspdfkit/shape/ellipse";
  cloudyBorderIntensity: number | null;
  cloudyBorderInset: InsetJSON | null;
  measurementBBox: IRectJSON | null;
};
type LineAnnotationJSON = ShapeAnnotationJSON & {
  type: "pspdfkit/shape/line";
  startPoint: [number, number];
  endPoint: [number, number];
  lineCaps?: LineCapsType | null;
  lines?: {
    points: [number, number][][];
    intensities: number[][];
  };
};
type PolygonAnnotationJSON = ShapeAnnotationJSON & {
  type: "pspdfkit/shape/polygon";
  points: [number, number][];
  cloudyBorderIntensity: number | null;
  lines?: {
    points: [number, number][][];
    intensities: number[][];
  };
};
type PolylineAnnotationJSON = ShapeAnnotationJSON & {
  type: "pspdfkit/shape/polyline";
  points: [number, number][];
  lineCaps?: LineCapsType | null;
  lines?: {
    points: [number, number][][];
    intensities: number[][];
  };
};
type RectangleAnnotationJSON = ShapeAnnotationJSON & {
  type: "pspdfkit/shape/rectangle";
  cloudyBorderIntensity: number | null;
  cloudyBorderInset?: InsetJSON | null;
  measurementBBox: IRectJSON | null;
};
type InkAnnotationJSON = BaseAnnotationJSON & {
  type: "pspdfkit/ink";
  lines: {
    points: [number, number][][];
    intensities: number[][];
  };
  lineWidth: number;
  strokeColor: string | null;
  backgroundColor: string | null;
  isDrawnNaturally: boolean;
  isSignature: boolean;
};
type LinkAnnotationJSON = BaseAnnotationJSON & {
  type: "pspdfkit/link";
  borderColor?: string | null;
  borderWidth?: number | null;
  borderStyle?: IBorderStyle | null;
};
type NoteAnnotationJSON = Omit<BaseAnnotationJSON, "type"> & {
  type: "pspdfkit/note";
  text?: {
    format: "plain";
    value: string;
  };
  icon?: string;
  color?: string;
};
type MediaAnnotationJSON = Omit<BaseAnnotationJSON, "type"> & {
  type: "pspdfkit/media";
  description: string | null;
  fileName: string | null;
  contentType: string | null;
  mediaAttachmentId: string | null;
};
type BaseTextMarkupAnnotationJSON = Omit<BaseAnnotationJSON, "type"> & {
  rects: [number, number, number, number][];
};
type TextMarkupAnnotationJSON = BaseTextMarkupAnnotationJSON & {
  type:
    | "pspdfkit/markup/highlight"
    | "pspdfkit/markup/squiggly"
    | "pspdfkit/markup/strikeout"
    | "pspdfkit/markup/underline"
    | "pspdfkit/markup/redaction";
  color: string | null;
};
type RedactionAnnotationJSON = BaseTextMarkupAnnotationJSON & {
  type: "pspdfkit/markup/redaction";
  fillColor?: string | null;
  outlineColor?: string | null;
  overlayText?: string | null;
  repeatOverlayText?: boolean | null;
  rotation?: number;
  color?: string | null;
};
type StampAnnotationJSON = Omit<BaseAnnotationJSON, "type"> & {
  type: "pspdfkit/stamp";
  stampType: StampKind;
  title: string | null;
  color?: string | null;
  subTitle?: string | null;
  subtitle: string | null;
  rotation: number | null;
  xfdfAppearanceStream?: string;
  xfdfAppearanceStreamOriginalPageRotation?: number;
  kind?: StampKind;
};
type TextAnnotationJSON = Omit<BaseAnnotationJSON, "type"> & {
  type: "pspdfkit/text";
  text: {
    format: "xhtml" | "plain";
    value: string;
  };
  fontColor?: string | null;
  backgroundColor?: string | null;
  font?: string | null;
  rotation?: number | null;
  fontSize?: number | null;
  fontStyle?: string[] | null;
  horizontalAlign?: "left" | "center" | "right";
  verticalAlign?: "top" | "center" | "bottom";
  callout?: {
    start: [number, number];
    knee?: [number, number] | null;
    end: [number, number];
    cap?: ILineCap | null;
    innerRectInset?: InsetJSON | null;
  } | null;
  borderStyle?: IBorderStyle | null;
  borderWidth?: number | null;
  borderColor?: string | null;
  isFitting?: boolean;
  lineHeightFactor?: number | null;
};
type UnknownAnnotationJSON = Omit<BaseAnnotationJSON, "type"> & {
  type: "pspdfkit/unknown";
};
type WidgetAnnotationJSON = Omit<BaseAnnotationJSON, "type"> & {
  type: "pspdfkit/widget";
  formFieldName: string;
  borderColor?: string | null;
  borderStyle?: IBorderStyle | null;
  borderDashArray?: number[] | null;
  borderWidth?: number | null;
  font?: string | null;
  fontSize?: "auto" | number | null;
  fontColor?: string | null;
  backgroundColor?: string | null;
  horizontalAlign?: "left" | "center" | "right" | null;
  verticalAlign?: "top" | "center" | "bottom" | null;
  fontStyle?: string[] | null | undefined;
  rotation?: number;
  additionalActions?: SerializedAdditionalActionsType | null;
  lineHeightFactor?: number | null;
};
type CommentMarkerAnnotationJSON = Omit<BaseAnnotationJSON, "type"> & {
  type: "pspdfkit/comment-marker";
};
export type AnnotationJSONUnion =
  | TextMarkupAnnotationJSON
  | TextAnnotationJSON
  | WidgetAnnotationJSON
  | RedactionAnnotationJSON
  | StampAnnotationJSON
  | NoteAnnotationJSON
  | LinkAnnotationJSON
  | InkAnnotationJSON
  | RectangleAnnotationJSON
  | PolylineAnnotationJSON
  | PolygonAnnotationJSON
  | LineAnnotationJSON
  | EllipseAnnotationJSON
  | ImageAnnotationJSON
  | UnknownAnnotationJSON
  | MediaAnnotationJSON
  | CommentMarkerAnnotationJSON;
// ...
type AnnotationsUnion = {
  [K in keyof AnnotationSerializerTypeMap]: AnnotationSerializerTypeMap[K]["annotation"];
}[keyof AnnotationSerializerTypeMap];
// ...
declare class ViewState extends ViewState_base {
  zoomIn(): ViewState;
  zoomOut(): ViewState;
  rotateLeft(): ViewState;
  rotateRight(): ViewState;
  goToNextPage(): ViewState;
  goToPreviousPage(): ViewState;
}

declare class InstantClient {
  clientId: string;
  userId: string | null | undefined;
}
// ...
type AnnotationsPressEvent = {
  annotation: AnnotationsUnion;
  nativeEvent: Event;
  preventDefault?: () => void;
  selected: boolean;
};

declare enum AnnotationsWillChangeReason {
  DRAW_START = "DRAW_START",
  DRAW_END = "DRAW_END",
  TEXT_EDIT_START = "TEXT_EDIT_START",
  TEXT_EDIT_END = "TEXT_EDIT_END",
  SELECT_START = "SELECT_START",
  SELECT_END = "SELECT_END",
  MOVE_START = "MOVE_START",
  MOVE_END = "MOVE_END",
  RESIZE_START = "RESIZE_START",
  RESIZE_END = "RESIZE_END",
  ROTATE_START = "ROTATE_START",
  ROTATE_END = "ROTATE_END",
  DELETE_START = "DELETE_START",
  DELETE_END = "DELETE_END",
  PROPERTY_CHANGE = "PROPERTY_CHANGE",
}

type PagePressEvent = {
  pageIndex: number;
  point: Point;
  nativeEvent: Event;
};

type AnnotationPresetsUpdateEvent = {
  preventDefault: () => boolean;
  currentPreset: AnnotationPresetID;
  currentPresetProperties: AnnotationPreset;
  newPresetProperties: AnnotationPreset;
};

type AnnotationsFocusEvent = {
  annotation: AnnotationsUnion;
  nativeEvent: FocusEvent;
};
type AnnotationsBlurEvent = {
  annotation: AnnotationsUnion;
  nativeEvent: FocusEvent;
};

type SaveStateChangeEvent = {
  hasUnsavedChanges: boolean;
};

type SearchTermChangeEvent = {
  term: string;
  preventDefault: () => void;
};

type TextLinePressEvent = {
  textLine: TextLine;
  point: Point;
  nativeEvent: Event;
};

type CropAreaChangeStartEvent = {
  cropBox: Rect;
  pageIndex: number;
};
type CropAreaChangeStopEvent = {
  cropBox: Rect;
  pageIndex: number;
};

type AnnotationsTransformEvent = {
  annotation: AnnotationsUnion;
};

type AnnotationsCopyEvent = {
  annotation: AnnotationsUnion;
};

type AnnotationsCutEvent = {
  annotation: AnnotationsUnion;
};

type AnnotationsPasteEvent = {
  annotation: AnnotationsUnion;
  formField?: FormField;
  previousAction: "COPY" | "CUT";
  originalAnnotation: AnnotationsUnion;
  originalFormField?: FormField;
};

type AnnotationsDuplicateEvent = {
  annotation: AnnotationsUnion;
  formField?: FormField;
  originalAnnotation: AnnotationsUnion;
  originalFormField?: FormField;
};
// ...
type Signature = InkAnnotation | ImageAnnotation;
interface HistoryEvent<T> {
  action: T;
  before: AnnotationsUnion;
  after: AnnotationsUnion;
}
export interface EventMap {
  "viewState.change": (
    viewState: ViewState,
    previousViewState: ViewState,
  ) => void;
  "viewState.currentPageIndex.change": (pageIndex: number) => void;
  "viewState.zoom.change": (zoom: number) => void;
  "annotationPresets.update": (event: AnnotationPresetsUpdateEvent) => void;
  "annotations.blur": (event: AnnotationsBlurEvent) => void;
  "annotations.change": () => void;
  "annotations.create": (annotations: List<AnnotationsUnion>) => void;
  "annotations.delete": (annotations: List<AnnotationsUnion>) => void;
  "annotations.didSave": () => void;
  "annotations.focus": (event: AnnotationsFocusEvent) => void;
  "annotations.load": (annotations: List<AnnotationsUnion>) => void;
  "annotations.press": (event: AnnotationsPressEvent) => void;
  "annotations.update": (annotations: List<AnnotationsUnion>) => void;
  "annotations.willChange": (event: {
    reason: AnnotationsWillChangeReason;
    annotations: List<AnnotationsUnion>;
  }) => void;
  "annotations.willSave": () => void;
  "annotationSelection.change": (annotation?: AnnotationsUnion) => void;
  "annotations.transform": (event: AnnotationsTransformEvent) => void;
  "annotations.copy": (event: AnnotationsCopyEvent) => void;
  "annotations.cut": (event: AnnotationsCutEvent) => void;
  "annotations.paste": (event: AnnotationsPasteEvent) => void;
  "annotations.duplicate": (event: AnnotationsDuplicateEvent) => void;
  "bookmarks.change": () => void;
  "bookmarks.create": (bookmarks: List<Bookmark>) => void;
  "bookmarks.update": (bookmarks: List<Bookmark>) => void;
  "bookmarks.delete": (bookmarks: List<Bookmark>) => void;
  "bookmarks.load": (bookmarks: List<Bookmark>) => void;
  "bookmarks.didSave": () => void;
  "bookmarks.willSave": () => void;
  "comments.change": () => void;
  "comments.create": (comments: List<Comment>) => void;
  "comments.delete": (comments: List<Comment>) => void;
  "comments.update": (comments: List<Comment>) => void;
  "comments.load": () => void;
  "comments.willSave": () => void;
  "comments.didSave": () => void;
  "instant.connectedClients.change": (
    clients: Map<string, InstantClient>,
  ) => void;
  "document.change": (operations: DocumentOperation[]) => void;
  "document.saveStateChange": (event: SaveStateChangeEvent) => void;
  "formFieldValues.update": (formFields: List<FormField>) => void;
  "formFieldValues.willSave": () => void;
  "formFieldValues.didSave": (res: {
    response: Response;
    error: Error;
  }) => void;
  "forms.willSubmit": (event: { preventDefault: () => void }) => void;
  "forms.didSubmit": (event: { preventDefault: () => void }) => void;
  "formFields.change": () => void;
  "formFields.create": (formFields: List<FormField>) => void;
  "formFields.delete": (formFields: List<FormField>) => void;
  "formFields.didSave": () => void;
  "formFields.load": (formFields: List<FormField>) => void;
  "formFields.update": (formFields: List<FormField>) => void;
  "formFields.willSave": () => void;
  "search.stateChange": (searchState: SearchState) => void;
  "search.termChange": (event: SearchTermChangeEvent) => void;
  "storedSignatures.change": () => void;
  "storedSignatures.create": (signature: Signature) => void;
  "storedSignatures.delete": (signature: Signature) => void;
  "storedSignatures.update": (signatures: List<Signature>) => void;
  "textLine.press": (event: TextLinePressEvent) => void;
  "textSelection.change": (selection: PublicTextSelection | null) => void;
  "history.change": (event: HistoryEvent<"undo" | "redo">) => void;
  "history.willChange": (event: {
    type: "create" | "update" | "delete";
    annotation: Annotation;
    preventDefault: () => void;
  }) => void;
  "history.clear": () => void;
  "history.redo": (event: HistoryEvent<"redo">) => void;
  "history.undo": (event: HistoryEvent<"undo">) => void;
  "page.press": (event: PagePressEvent) => void;
  "inkSignatures.create": (signature: Signature) => void;
  "inkSignatures.delete": (signature: Signature) => void;
  "inkSignatures.update": (signatures: Signature[]) => void;
  "inkSignatures.change": () => void;
  "cropArea.changeStart": (opts: CropAreaChangeStartEvent) => void;
  "cropArea.changeStop": (opts: CropAreaChangeStopEvent) => void;
  "documentComparisonUI.start": (opts: DocumentComparisonUIStartEvent) => void;
  "documentComparisonUI.end": () => void;
  "annotationNote.press": (event: AnnotationNotePressEvent) => void;
  "annotationNote.hover": (event: AnnotationNoteHoverEvent) => void;
  "comments.mention": (event: CommentsMentionEvent) => void;
}